feat!: Linked POS module to DB. Updated Stock module to work with DB. Other minor changes

+++ EDSRetail-Alpha/databaseAPI/Models/Sales.cs
@@ -6,8 +6,8 @@
         public string SaleID { get; set; }
         public string SaleIDHASH { get; set; }
         public string SKU { get; set; }
-        public float QTY { get; set; }
-        public float Price { get; set; }
+        public double QTY { get; set; }
+        public double Price { get; set; }
         public string Description { get; set; }
 
     }
