+++ EDSRetail-Alpha/databaseAPI/Models/Stock.cs
@@ -6,9 +6,9 @@
         public string SKU { get; set; }
         public string Barcode { get; set; }
         public string Description { get; set; }
-        public float QTYOnHand { get; set; }
-        public float SellPrice { get; set; }
-        public float CostPrice { get; set; }
+        public double QTYOnHand { get; set; }
+        public double SellPrice { get; set; }
+        public double CostPrice { get; set; }
 
         public bool CanTrade { get; set; }
         public bool ScaleItem { get; set; }
